+++ core/src/com/unciv/logic/automation/civilization/BarbarianAutomation.kt
@@ -1,8 +1,8 @@
 package com.unciv.logic.automation.civilization
 
 import com.unciv.Constants
-import com.unciv.logic.automation.BattleHelper
-import com.unciv.logic.automation.units.UnitAutomation
+import com.unciv.logic.automation.unit.BattleHelper
+import com.unciv.logic.automation.unit.UnitAutomation
 import com.unciv.logic.civilization.CivilizationInfo
 import com.unciv.logic.map.MapUnit
